Student to Course table relation. one to many relation

+++ src/main/java/com/vidyasagar/attendance/api/v1/service/impl/CourseServiceImpl.java
@@ -1,7 +1,11 @@
 package com.vidyasagar.attendance.api.v1.service.impl;
 
+import com.vidyasagar.attendance.api.v1.dto.response.CourseDTO;
+import com.vidyasagar.attendance.api.v1.mapper.CourseMapper;
+import com.vidyasagar.attendance.api.v1.repository.StudentRepository;
 import com.vidyasagar.attendance.api.v1.service.CourseService;
 import com.vidyasagar.attendance.entity.Course;
+import com.vidyasagar.attendance.entity.Student;
 import com.vidyasagar.attendance.exception.ResourceNotFoundException;
 import com.vidyasagar.attendance.api.v1.repository.CourseRepository;
 import org.springframework.stereotype.Service;
@@ -11,10 +15,18 @@ import java.util.List;
 @Service
 public class CourseServiceImpl implements CourseService {
     private final CourseRepository courseRepository;
+    private final StudentRepository studentRepository;
+    private final CourseMapper courseMapper;
 
     // Constructor Injection
-    public CourseServiceImpl(CourseRepository courseRepository) {
+    public CourseServiceImpl(
+            CourseRepository courseRepository,
+            StudentRepository studentRepository,
+            CourseMapper courseMapper
+    ) {
         this.courseRepository = courseRepository;
+        this.studentRepository = studentRepository;
+        this.courseMapper = courseMapper;
     }
 
     @Override
@@ -23,37 +35,55 @@ public class CourseServiceImpl implements CourseService {
     }
 
     @Override
-    public Course saveCourse(Course course) {
-        return courseRepository.save(course);
+    public CourseDTO saveCourse(CourseDTO courseDTO) {
+        Course course = courseMapper.toEntity(courseDTO);
+
+        // If student is attached, ensure it's loaded from DB(best practice)
+        if(courseDTO.getStudentId() != null) {
+            Student student = studentRepository.findById(courseDTO.getStudentId())
+                    .orElseThrow(() -> new ResourceNotFoundException("Student not found"));
+            course.setStudent(student);
+        }
+
+        Course saved = courseRepository.save(course);
+        return courseMapper.toDTO(saved);
     }
 
     @Override
-    public List<Course> getAllCourses() {
-        return courseRepository.findAll();
+    public List<CourseDTO> getAllCourses() {
+        return courseMapper.toDTOList(courseRepository.findAll());
     }
 
     @Override
-    public Course getCourseById(Long id) {
-        return courseRepository.findById(id)
+    public CourseDTO getCourseById(Long id) {
+        Course course = courseRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Course not found with id " + id));
+        return courseMapper.toDTO(course);
     }
 
     @Override
-    public Course updateCourse(Long id, Course courseDetails){
+    public CourseDTO updateCourse(Long id, CourseDTO courseDetails){
         Course course = courseRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Course not found with id " + id));
 
         course.setCredits(courseDetails.getCredits());
         course.setTitle(courseDetails.getTitle());
 
-        return courseRepository.save(course);
+        if(courseDetails.getStudentId() != null) {
+            Student student = studentRepository.findById(id)
+                    .orElseThrow(() -> new ResourceNotFoundException("Student not found"));
+            course.setStudent(student);
+        }
+
+        Course updated = courseRepository.save(course);
+        return courseMapper.toDTO(updated);
     }
 
     @Override
     public void  deleteCourse(Long id) {
-        Course course = courseRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Course not found with id " + id));
-
+        if(!courseRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Course not found with id " + id);
+        }
         courseRepository.deleteById(id);
     }
 }
